Do not hard code absolute developer paths
By default, Composer creates a classmap of PHP classes
to source files. Where those source files are written
into the application or repository, the Composer
classmap will hard code absolute paths to those
files within the classmap.

In situations where a phar file is created on a
development machine and then deployed to a live
environment this is unhelpful because the absolute
paths will not exist where the vendor.phar is loaded.

This commit re-writes the autoload files generated
by Composer to replace the absolute paths with a
common path to wp-content which is used in many
hosting environments:

    /var/www/html

For users who prefer not to have this, the
PHAR_INSTALL_PATH_TO_WP_CONTNET environment variable
can be used to change the path.

See:
    https://getcomposer.org/doc/articles/autoloader-optimization.md